--- src/services/driverServer/src/main/java/org/example/driverserver/exception/DuplicateFieldException.java
@@ -1,11 +1,0 @@
-package org.example.driverserver.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class DuplicateFieldException extends RuntimeException {
-    private String message;
-    private Long errorCode;
-}
