--- src/services/driverServer/src/main/java/org/example/driverserver/exception/ErrorMessage.java
@@ -1,13 +1,0 @@
-package org.example.driverserver.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorMessage {
-    private Long errorCode;
-    private String errorMessage;
-}
