--- src/services/driverServer/src/main/java/org/example/driverserver/exception/NotFoundException.java
@@ -1,11 +1,0 @@
-package org.example.driverserver.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class NotFoundException extends RuntimeException{
-    private String message;
-    private Long errorCode;
-}
