+++ src/services/driverService/src/main/java/org/example/driverservice/model/Driver.java
@@ -1,0 +1,35 @@
+package org.example.driverservice.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Driver {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(unique = true, nullable = false)
+    private String email;
+
+    @Column(nullable = false)
+    private String phone;
+
+    @OneToMany(mappedBy = "driver")
+    private List<Car> cars;
+
+    @Column(nullable = false)
+    private boolean deleted = Boolean.FALSE;
+}
